+++ javascript/node-postgres/package.json
@@ -13,8 +13,7 @@
     "@aws-sdk/dsql-signer": "^3.705.0",
     "assert": "2.1.0",
     "pg": "^8.13.1",
-    "uuid": "^11.0.2",
-    "cross-spawn": "7.0.6"
+    "uuid": "^11.0.2"
   },
   "devDependencies": {
     "jest": "^29.7.0"
