+++ javascript/node-postgres/test/smoke.test.js
@@ -1,6 +1,6 @@
 import { example } from '../src/index.js';
 
 test('Smoke test', async () => {
-  await example(process.env.CLUSTER_ENDPOINT, process.env.REGION);
+  await example();
   return Promise.resolve();
 });
