--- javascript/nodejs/README.md
@@ -1,32 +1,0 @@
-# Aurora DSQL Node-Postgres code examples
-
-## Overview
-
-The code examples in this topic show you how to use the Node-Postgres work with Aurora DSQL. 
-
-## Run the examples
-
-### Prerequisites
-
-* node version >=18.0 is needed
-
-### Setup test running environment 
-
-```sh
-npm install
-```
-
-### Run the example tests
-
-```sh
-# Use the account credentials dedicated for javascript
-export CLUSTER_ENDPOINT="<your cluster endpoint>"
-export REGION="<your cluster region>"
-npm test
-```
-
----
-
-Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved. 
-
-SPDX-License-Identifier: MIT-0
