--- javascript/nodejs/src/index.js
@@ -1,64 +1,0 @@
-import { DsqlSigner } from "@aws-sdk/dsql-signer";
-import pg from "pg";
-import assert from "node:assert";
-const { Client } = pg;
-
-async function example(clusterEndpoint, region) {
-  let client;
-  try {
-    // The token expiration time is optional, and the default value 900 seconds
-    const signer = new DsqlSigner({
-      hostname: clusterEndpoint,
-      region,
-    });
-    const token = await signer.getDbConnectAdminAuthToken();
-    // <https://node-postgres.com/apis/client>
-    // By default `rejectUnauthorized` is true in TLS options
-    // <https://nodejs.org/api/tls.html#tls_tls_connect_options_callback>
-    // The config does not offer any specific parameter to set sslmode to verify-full
-    // Settings are controlled either via connection string or by setting
-    // rejectUnauthorized to false in ssl options
-    client = new Client({
-      host: clusterEndpoint,
-      user: "admin",
-      password: token,
-      database: "postgres",
-      port: 5432,
-      // <https://node-postgres.com/announcements> for version 8.0
-      ssl: true,
-      rejectUnauthorized: false
-    });
-
-    // Connect
-    await client.connect();
-
-    // Create a new table
-    await client.query(`CREATE TABLE IF NOT EXISTS owner (
-      id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      name VARCHAR(30) NOT NULL,
-      city VARCHAR(80) NOT NULL,
-      telephone VARCHAR(20)
-    )`);
-
-    // Insert some data
-    await client.query("INSERT INTO owner(name, city, telephone) VALUES($1, $2, $3)", 
-      ["John Doe", "Anytown", "555-555-1900"]
-    );
-
-    // Check that data is inserted by reading it back
-    const result = await client.query("SELECT id, city FROM owner where name='John Doe'");
-    assert.deepEqual(result.rows[0].city, "Anytown")
-    assert.notEqual(result.rows[0].id, null)
-
-    await client.query("DELETE FROM owner where name='John Doe'");
-
-  } catch (error) {
-    console.error(error);
-    raise
-  } finally {
-    client?.end()
-  }
-  Promise.resolve()
-}
-
-export { example }
